test: fix test authz and add authz_execute
