+++ integration_tests/test.sh
@@ -2,6 +2,7 @@
 pytest -m normal -vv
 pytest -m gov -vv
 pytest -m authz -vv
+pytest -m authz_execute -vv
 pytest -m staking -vv
 pytest -m vesting -vv
 pytest -m byzantine -vv
